Add Control Functionality
The main controls in the CommandGrid (Hotkey textboxes, Name textboxes, File
textblocks, and Clear buttons) now have user functionality and save to database
more easily. Hotkey textboxes have the most processing because they only allow
valid inputs from the user and facilitate the user's experience with them.

However, there remain issues with hotkey input. The ALT key modifier is not
recognized because I could not figure out how to properly stop it from being
recognized as a System key. There may also be some keys not yet accounted for.

In the future, I would like to abstract my code better so that not every control
method is under MainWindow.xaml.cs, if possible. This also goes for the
XAML where the large amounts of controls and properties makes the code lengthy.
I believe it is possible to create a resource file where I can make a style for
each control similar to CSS class styling.

- MainWindow.xaml
    > Add multiple functions and properties to Hotkey and Name textboxes,
        File textblocks, and clear buttonns.
- DataAccess.cs
    > Add method to save a singular command (Hotkey, Name, and File_Path).
- MainWindow.xaml.cs
    > Add functionality to Hotkey textboxes.
         * Only allow valid keys for hotkey input.
         * Facilitate hotkey input by showing certain text (like "SHIFT + " upon
            pressing SHIFT).
         * Save hotkey text to database upon changing to a valid input.
    > Add functionality to Name textboxes
         * Save custom audio name to database upon pressing the Enter button.
         * Limit length of name.
    > Add functionality to File textblocks
         * Enabled drag and drop on textblock.
         * Dropping a valid file-type on a File textblock will show the entire
            file path in the textblock and show the shortened file name in
            the corresponding Name textbox.
         * Hovering over the textblock will show the entire file path because
            the entire file path may not always fit visually in the textblock.
    > Add functionality to Clear and Clear All buttons
         * Clicking the Clear button will clear the text in the Name textbox and
            the File textblock of the corresponding row. Saves command after
           clearing.
         * Clicking the Clear All button will clear the text in all of the Name
            textboxes and File textblocks in the profile, then save the cleared
            data.