Global Hotkeys Play Audio On Command
The new GlobalHotkey class takes the hotkey that the user enters and registers it on a global level to play the audio file linked to the command.
Now, the user can enter custom hotkeys that appropriately register, unregister, and re-register to play local audio files anywhere on their computer.

There are draw-backs to having global hotkeys. If the user creates a global hotkey that is also used by Windows or another program for other
purposes, then the other functionality will be cancelled out and the attached audio file will play instead. Some of these cases have already been
accounted for, and I intend to account for more cases (no letting the user have ALT + F4 as a hotkey, for example).

A new table exists in the database to hold the key and key modifier data (stored as integers, converted to Key and KeyModifier objects upon load).
To support this, DataAccess.cs has new methods to save and load GlobalHotkey data. On application start, the global hotkeys are loaded and
registered. On application close, the global hotkeys are unregistered and disposed of.

Removed Tab as a potential hotkey prefix/modifier. Restored Alt key as a potrntial hotkey prefix/modifier.