Set Default Hotkeys in Database
The database is now stored with a set of hotkeys by default (including virtual key codes and modifiers). This way, the user has a set of hotkeys to
play with right out the gate. The hotkeys themselves use the CTRL + SHIFT modifiers and all integer keys
(0-9) as well as alphabetical keys (A-Z), though they are listed in accordance with standard QWERTY keyboards.

It's unlikely that this is the ideal set of default hotkeys, but here are the reasons behind my choices:
- The CTRL + SHIFT modifiers give minimal annoyance when used in a global setting because they don't
   cancel out many commonly used hotkeys.
- Only integers and alphabetical keys were chosen for memorability, not comfort. The user might more
   easily associate a sound effect with a letter or number than a comma, or a forward slash.