Add "Deactivate Hotkeys" CheckBox and Layout Changes
The project has been renamed from Soundboard Hotkeys to Hotkey Soundboard.

The new "Deactivate Hotkeys" CheckBox allows for hotkeys to be disabled while the checkbox is checked and re-enabled when it's unchecked. By default the checkbox is unchecked because hotkeys should be activated by default.

While hotkeys are deactivated, the user can still switch between profiles. The global hotkey and command data continues to be loaded from the database, but the global hotkeys will not register. Deactivating hotkeys will allow for unaltered hotkey usage elsewhere on their desktop without having to close this program.

Some changes were made to the layout of MainWindow.
     > Row containing the column identifying labels and Clear All button have been moved out of
        CommandGrid and into the overall grid. This keeps the labels and button visible at all times
        while scrolling through commands.
     > The context menus of the Hotkey and Name textboxes
        have been removed.