Create README.md and Create Published Version
A version of the application has been created for non-developer use in a self-contained, single-file (ProfileDB excluded) executable.
In the ReadMe, a link is provided to download the application. Some notes about how I would like to expand the project are given in 
the "Developer Notes" section at the bottom of the ReadMe.